[refactor]將 py 的 api 往外抽，並嘗試使用 init
