+++ api/__init__.py
@@ -1,0 +1,1 @@
+from .home import router as home_router
